feat(types): export RefreshType
