+++ lib/utils/token/refreshToken.ts
@@ -6,19 +6,7 @@ import {
 } from "../../sessionManager";
 import { isCustomDomain, sanitizeUrl } from "..";
 import { clearRefreshTimer, setRefreshTimer } from "../refreshTimer";
-
-export interface RefreshTokenResult {
-  success: boolean;
-  error?: string;
-  [StorageKeys.accessToken]?: string;
-  [StorageKeys.idToken]?: string;
-  [StorageKeys.refreshToken]?: string;
-}
-
-export enum RefreshType {
-  refreshToken,
-  cookie,
-}
+import { RefreshTokenResult, RefreshType } from "../../main";
 
 /**
  * refreshes the token
